Any LaTeX math fragments in an item will now be rendered inline (if possible) during drill sessions. This means org-drill can now be used to memorize maths equations. LaTeX fragments work correctly if placed inside cloze text, i.e. between square brackets. I suggest using \(…\) to delimit fragments, as using \[…\] may be difficult for org-drill to differentiate from cloze text.
'Cloze' overlays ([…]) are now very high priority, to ensure that any inner
overlays (such as inline LaTeX images) are not displayed during presentation
of the item.

Where dates are stored in item properties, we now use inactive rather than
active timestamps.

New command 'org-drill-relearn-item' lets you reset the repetition count for
an item, as if it was failed, without actually recording it as failed.